requirements.txt
aimlsse-api>=0.3.17
fastapi>=0.88.0
Fiona>=1.8.22
geopandas>=0.12.2
PyYAML>=6.0
requests>=2.28.1
uvicorn>=0.20.0
schedule>=1.1.0
sentinelsat>=1.1.1